scripts/audio2face_3d_microservices_interaction_app/requirements.txt
numpy==1.26.4
scipy==1.13.0
grpcio==1.64.0
protobuf==4.24.1
PyYAML==6.0.1
pandas==2.2.2
matplotlib==3.9.1
fastapi
uvicorn
google-cloud-storage
